fix(creditos): soportar materias con créditos decimales (0.5)
Banner publica materias con créditos fraccionarios (ej. "Seminario De Desarrollo Personal", 0.5). La columna Materia.Creditos era INTEGER: Postgres redondeaba al insertar (0.5 -> 0) sin lanzar error, así que esas materias aparecían con 0 créditos, no sumaban al contador y no consumían cupo del límite de 20.

Toda la cadena asumía enteros, así que el decimal se propaga capa por capa:
- BD: Materia.Creditos pasa a NUMERIC(4,2). Como init.sql solo corre con el volumen vacío, se añade migrar_esquema.py (migraciones idempotentes) que el ETL ejecuta en cada corrida y aplica el ALTER TABLE en bases ya creadas.
- ETL: parser.obtener_creditos() conserva el decimal de Banner.
- Backend: credits es float en los modelos; el repositorio convierte el Decimal de psycopg (no se puede sumar con floats) a float.
- Generador: la poda por créditos usa float con tolerancia EPSILON, para que una suma que da justo el tope (19.5 + 0.5 = 20) no se descarte por error de coma flotante.
- Frontend: credits es double en los tres modelos; credit_utils.dart centraliza parseo, redondeo del acumulado y formato (0.5, 3, sin .0 sobrante).

De paso, dos arreglos que salieron del camino:
- main.py inyectaba el tope del cliente como 'credit_limit', pero el generador lo lee de 'max_credits': el límite enviado por el frontend nunca se aplicaba (caía siempre al default de 20 del generador). Pasaba desapercibido porque el frontend envía justo 20.
- tests/test_correctness.py no compilaba: su factory no pasaba campus/seatsAvailable/seatsMaximum, que el modelo exige desde que se añadieron.

Verificado end-to-end contra la base local: tras el ETL, las 2 materias de 0.5 se guardan como 0.50, /api/subjects las expone como 0.5 y el generador las incluye respetando el tope (creditLimit 0.4 poda, 0.5 cabe).

+++ backend/app/models.py
@@ -19,7 +19,8 @@ class ClassOption(BaseModel):
     
     nrc: str
     group_id: int = Field(..., alias='groupId')
-    credits: int
+    # Decimal: hay materias de créditos fraccionarios (ej. 0.5) en Banner.
+    credits: float
     campus: str
     seats_available: int = Field(..., alias='seatsAvailable')
     seats_maximum: int = Field(..., alias='seatsMaximum')
@@ -31,7 +32,7 @@ class Subject(BaseModel):
 
     code: str
     name: str
-    credits: int
+    credits: float
     class_options: List[ClassOption] = Field(..., alias='classOptions')
 
 
@@ -47,7 +48,8 @@ class GenerateScheduleRequest(BaseModel):
     subjects: List[SubjectIdentifier]
     filters: Dict[str, Any]
     # El frontend envía 'creditLimit' (camelCase), se usa un alias para que Pydantic lo entienda como 'credit_limit'.
-    credit_limit: int = Field(..., alias='creditLimit')
+    # Float porque las materias pueden sumar medios créditos (ej. 19.5).
+    credit_limit: float = Field(..., alias='creditLimit')
     # El cliente declara si es móvil; el backend limita los resultados solo en
     # ese caso (evita agotar la memoria del navegador móvil). Default false para
     # compatibilidad con clientes que no lo envíen.
